Custom Tab Filter Fixed
